Fix bug in pagination URLs
The paginator used the site BaseURL and concatenated it with the
relative URL of the page. This breaks if the site is not hosted on a
top-level domain.

For example, site hosted at example.com/foobar/ would have pagination
links as example.com/foobar/foobar/post/page/2 instead of the correct
example.com/foobar/post/page/2. Now, the link is /foobar/post/page/2,
which should be interpreted correctly by the browser.